Added `cargo_manifest_dir` submodule to `cargo_build_script_runner`.
